fixed proposal integration
